Move modules to app.
